update settings + create engine canvas file
