+++ iDibujo/iDibujo/Util/Extensions.swift
@@ -7,6 +7,25 @@
 //
 
 import UIKit
+import SwiftUI
+
+// MARK: - The extension provides a method to capture a snapshot of the displayed element or the view itself.
+extension View {
+    func snapshot() -> UIImage {
+        let controller = UIHostingController(rootView: self)
+        let view = controller.view
+
+        let targetSize = controller.view.intrinsicContentSize
+        view?.bounds = CGRect(origin: .zero, size: targetSize)
+        view?.backgroundColor = .clear
+
+        let renderer = UIGraphicsImageRenderer(size: targetSize)
+
+        return renderer.image { _ in
+            view?.drawHierarchy(in: controller.view.bounds, afterScreenUpdates: true)
+        }
+    }
+}
 
 // MARK: - Custom button for each BUTTON color
 class CustomButton: UIButton {
